Add default create-react-app .gitignore
This is the .gitignore file that create-react-app generates by default.
It tells Git not to track certain files that are either irrelevant or
that aren't often tracked in source control for some other reason.

This commit also removes .DS_Store.

The default .gitignore includes an entry for node_modules, but it was
removed for this commit because node_modules may or may not be ignored
depending on what decision is reached in this issue:

https://github.com/mozilla/mdv2/issues/1